requirements.txt
Flask==3.1.0
Pillow==11.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
PyYAML==6.0.2
